refactor(stores): Implement Autobuy, AutoUpgrade and RealmManager managers.
