--- src/lib/helpers/realmManager.ts
@@ -1,114 +1,0 @@
-import { derived, writable, type Readable, type Writable } from 'svelte/store';
-import { statManager, purpleRealmUnlocked } from '$stores/gameStore';
-import { CURRENCIES, CurrenciesTypes } from '$data/currencies';
-import { STATS } from '$helpers/statConstants';
-import type { Currency } from '$lib/types';
-import type { Component } from 'svelte';
-import AtomRealm from '@components/prestige/AtomRealm.svelte';
-import PhotonRealm from '@components/prestige/PhotonRealm.svelte';
-
-export interface RealmConfig {
-	component: Component;
-	store: Writable<number>;
-	activeClasses: string;
-	currency: Currency;
-	footerTheme: RealmFooterTheme;
-	id: string;
-	isUnlocked: () => boolean;
-	props?: Record<string, unknown>;
-	title: string;
-}
-
-export interface RealmFooterTheme {
-	baseClass: string;
-	hoverClass: string;
-	mutedClass: string;
-	mutedHoverClass: string;
-}
-
-interface RealmManagerStore {
-	availableRealms: RealmConfig[];
-	realmValues: Record<string, number>;
-	selectedRealm: string;
-}
-
-class RealmManager implements Readable<RealmManagerStore> {
-	public readonly realms: RealmConfig[];
-	private readonly _selectedRealm: Writable<string>;
-	private readonly _availableRealms: Readable<RealmConfig[]>;
-	private readonly _realmValues: Readable<Record<string, number>>;
-	public readonly subscribe: (
-		this: void,
-		run: (value: RealmManagerStore) => void,
-		invalidate?: (value?: RealmManagerStore) => void
-	) => () => void;
-
-	constructor() {
-		this.realms = [
-			{
-				activeClasses: 'bg-accent-500/60 border-accent-400/50',
-				component: AtomRealm,
-				currency: CURRENCIES[CurrenciesTypes.ATOMS],
-				footerTheme: {
-					baseClass: 'text-white/60',
-					hoverClass: 'hover:text-white',
-					mutedClass: 'text-white/40',
-					mutedHoverClass: 'hover:text-white'
-				},
-				id: 'atoms',
-				store: statManager.getNumber(STATS.ATOMS)!.store,
-				title: 'Atoms Realm',
-				isUnlocked: () => true
-			},
-			{
-				activeClasses: 'bg-realm-500/60 border-realm-400/50',
-				component: PhotonRealm,
-				currency: CURRENCIES[CurrenciesTypes.PHOTONS],
-				footerTheme: {
-					baseClass: 'text-white/85',
-					hoverClass: 'hover:text-purple-200',
-					mutedClass: 'text-white/70',
-					mutedHoverClass: 'hover:text-purple-200'
-				},
-				id: 'photons',
-				store: statManager.getNumber(STATS.PHOTONS)!.store,
-				title: 'Purple Realm',
-				isUnlocked: () => purpleRealmUnlocked.get()
-			}
-		];
-
-		this._selectedRealm = writable(this.realms[0].id);
-
-		this._availableRealms = derived(purpleRealmUnlocked, ($purpleRealmUnlocked) => {
-			return this.realms.filter((r) => r.isUnlocked());
-		});
-
-		const storesToWatch = this.realms.map((r) => r.store);
-		this._realmValues = derived(storesToWatch, ($values) => {
-			return this.realms.reduce(
-				(acc, screen, index) => {
-					acc[screen.id] = $values[index];
-					return acc;
-				},
-				{} as Record<string, number>
-			);
-		});
-
-		const combinedStore = derived(
-			[this._selectedRealm, this._availableRealms, this._realmValues],
-			([$selectedRealm, $availableRealms, $realmValues]) => ({
-				selectedRealm: $selectedRealm,
-				availableRealms: $availableRealms,
-				realmValues: $realmValues
-			})
-		);
-
-		this.subscribe = combinedStore.subscribe;
-	}
-
-	selectRealm(realmId: string) {
-		this._selectedRealm.set(realmId);
-	}
-}
-
-export const realmManager = new RealmManager();
